design: change design to allow multiple implementation
